Merged PR 26: Removed ObjectId from SecurityKey
Removed ObjectId from the SecurityKey model.

PROTECT:
Protect endpoint now only expects an item to encrypt. It will thereafter respond back with the encrypted item and the id of the SecurityKey (which holds the decryption key for the item in the database).
Example of request:
.-------------------------------------------------------------------.
{
  "plainText": "string"
}
Output:
{
  "securityKeyId": "cf4b2fbb-6217-4e83-9739-cc1f2dfdc683",
  "protectedText": "SlRVBG0f8J4E7yU3MkvYkLw2ZyTKCgpabiTdtiCBRxs="
}
.-------------------------------------------------------------------.
.
.
REPROTECT:
Reprotect endpoint now expects a SecurityKey id and the protected item. It will then respond back with the encrypted item (in new form) and the id of the SecurityKey.
Example of request:
.-------------------------------------------------------------------.
{
  "securityKeyId": "cf4b2fbb-6217-4e83-9739-cc1f2dfdc683",
  "protectedText": "SlRVBG0f8J4E7yU3MkvYkLw2ZyTKCgpabiTdtiCBRxs="
}
Output:
{
  "securityKeyId": "cf4b2fbb-6217-4e83-9739-cc1f2dfdc683",
  "protectedText": "/kehWqKA4NbeQlResqpiS2MBPz4wQE6ShTGNM6yC3q8="
}
.-------------------------------------------------------------------.
.
.
UNPROTECT:
Unprotect endpoint now also expects a SecurityKey id and the protected item. It will then respond with the decrypted item in plain text.
Example of request:
.-------------------------------------------------------------------.
{
  "securityKeyId": "cf4b2fbb-6217-4e83-9739-cc1f2dfdc683",
  "protectedText": "/kehWqKA4NbeQlResqpiS2MBPz4wQE6ShTGNM6yC3q8="
}
Output:
{
  "unprotectedText": "string"
}
.-------------------------------------------------------------------.
.
.
DELETE:
Delete endpoint now only expects the id of the security key.
Example of request:
.-------------------------------------------------------------------.
{
  "securityKeyId": "cf4b2fbb-6217-4e83-9739-cc1f2dfdc683"
}
Output:
DELETED
.-------------------------------------------------------------------.

Related work items: #60, #61, #62, #63, #64, #65